Add files via upload
There are a number of similarly formatted files across year folders and month subfolders. I received a request to pull the number of accounts (rows) in the files each month and the percentage of those accounts that were marked to be extended. There were a number of different naming conventions within folders and the files themselves. The code accounts for the following issues:
1. Differences in subfolder names (i.e. 01 - January vs. 01- January)
2. Files that contained multiple sheets with the same columns (I pulled whichever had the largest number of "Extend" values)